Pring board while playing (#4)(#1)
Added board serializer and now print the board while running the trials to give feedback about what is happening.
Avoid NoSuchElementException in ZeroTreenode when the branch does not contain the move (a pass or resign).
In ZeroAgent.selectBranch return pass if there are no valid moves.
Use "diagonals" instead of "corners" to refer to the diagonal placements from a point on the grid.
Add GoString tests.