Upgrade libraries, use GPU, fix leak (#22)(#28)(#29)
Fixed several memory leaks.
No more memory leak when running without GPU on nd4j 1.0.3b, but
there still seems to be a leak using 1.0.4b and 1.0.6b of Nd4j or if using GPU.
Add more entries in performance csv.
Add ObjectSizer to help with leak debugging.
It can be used to determine if objects are growing in size unreasonably.